+++ src/views/Benchmark.vue
@@ -1,10 +1,11 @@
+<!-- eslint-disable vue/multi-word-component-names -->
 <template>
 	<!-- Modal area -->
 	<SettingsModal />
 
   <!-- App area -->
   <div class="h-screen overflow-hidden scrollbar-hide">
-    <NavbarComponent />
+    <Navbar />
     <BenchmarkGrid />
     <div 
       class="relative w-full h-[55vh] overflow-auto bg-white p-4 shadow-md rounded-md"
@@ -17,9 +18,9 @@
   </div>
 </template>
 
-<script lang="ts">
+<script setup lang="ts">
 import { watch } from 'vue'
-import NavbarComponent from '@/components/Navbar.vue'
+import Navbar from '@/components/Misc/Navbar.vue'
 import BenchmarkGrid from '@/components/Benchmark/BenchmarkGrid.vue'
 import SpeckleViewer from '@/components/ModelViewer/SpeckleViewer.vue'
 
@@ -33,27 +34,13 @@ import SettingsModal from '@/components/Modals/SettingsModal.vue'
  * Dashboard view.
  * This component represents the main dashboard view of the application.
  */
-export default {
-	name: 'DashboardView',
-	components: {
-		NavbarComponent,
-    BenchmarkGrid,
-		SettingsModal,
-    SpeckleViewer
-	},
-	setup() {
-		//Load materials from the store on startup
-		const navStore = useNavigationStore()
-		
-		// Watch for changes in the active page and update viewer colors
-		// TODO: Have this in the navStore instead?
-		watch(() => navStore.activePage , (newVal) => {
-			if(newVal === 'Benchmark') {
-				console.log("Benchmark page - No graphics to update")
-			}
-		})
+const navStore = useNavigationStore()
 
-		return {}
+// Watch for changes in the active page and update viewer colors
+// TODO: Have this in the navStore instead?
+watch(() => navStore.activePage , (newVal) => {
+	if(newVal === 'Benchmark') {
+		console.log("Benchmark page - No graphics to update")
 	}
-}
+})
 </script>
