+++ src/views/Dashboard.vue
@@ -1,3 +1,4 @@
+<!-- eslint-disable vue/multi-word-component-names -->
 <template>
 	<!-- Modal area -->
 	<NewGroupModal />
@@ -8,7 +9,7 @@
 
   <!-- App area -->
 	<div class="flex">
-		<NavbarComponent />
+		<Navbar />
 		<Sidebar />
 		<div 
 			class="relative inset-y-16 w-full h-[calc(100vh-4rem)] bg-gray-100 overflow-auto" 
@@ -26,12 +27,12 @@
 	</div>
 </template>
 
-<script lang="ts">
+<script setup lang="ts">
 import { watch } from 'vue'
 import Sidebar from '@/components/Sidebar/Sidebar.vue'
 import Slideover from '@/components/SlideOver/Sliderover.vue'
 import SpeckleViewer from '@/components/ModelViewer/SpeckleViewer.vue'
-import NavbarComponent from '@/components/Navbar.vue'
+import Navbar from '@/components/Misc/Navbar.vue'
 import { useProjectStore } from '@/stores/main'
 import { useNavigationStore } from '@/stores/navigation'
 import { useSpeckleStore } from '@/stores/speckle'
@@ -58,54 +59,35 @@ import SettingsModal from '@/components/Modals/SettingsModal.vue'
  * Dashboard view.
  * This component represents the main dashboard view of the application.
  */
-export default {
-	name: 'DashboardView',
-	components: {
-		NavbarComponent,
-		SpeckleViewer,
-		Sidebar,
-		Slideover,
-		NewGroupModal,
-		MaterialMappingModal,
-		SaveFilterModal, 
-		AssemblyModal,
-		SettingsModal,
-	},
-	setup() {
-		//Load materials from the store on startup
-		const navStore = useNavigationStore()
-		const projectStore = useProjectStore()
-		const speckleStore = useSpeckleStore()
+const navStore = useNavigationStore()
+const projectStore = useProjectStore()
+const speckleStore = useSpeckleStore()
 
-		// Preload all needed resources
-		preloadDashboardData()
+// Preload all needed resources
+preloadDashboardData()
 
-		// Watch for changes in the active page and update viewer colors
-		// TODO: Have this in the navStore instead?
-		watch(() => navStore.activePage , (newVal) => {
-			if(newVal === 'Overview') {
-				updateProjectGroups(true)
-				const tree =	projectStore.getGroupTree()?.children
-				speckleStore.calculateGroupColors(tree)
-			} else if(newVal === 'Mapping') {
-				const mappingColors = setMappingColorGroup()
-				speckleStore.setColorGroups(mappingColors)
-			} else if(newVal === 'Results') {
-				// Trigger calc for project
-				const calculator = new EmissionCalculator()
-				calculator.calculateEmissions()
+// Watch for changes in the active page and update viewer colors
+// TODO: Have this in the navStore instead?
+watch(() => navStore.activePage , (newVal) => {
+	if(newVal === 'Overview') {
+		updateProjectGroups(true)
+		const tree =	projectStore.getGroupTree()?.children
+		speckleStore.calculateGroupColors(tree)
+	} else if(newVal === 'Mapping') {
+		const mappingColors = setMappingColorGroup()
+		speckleStore.setColorGroups(mappingColors)
+	} else if(newVal === 'Results') {
+		// Trigger calc for project
+		const calculator = new EmissionCalculator()
+		calculator.calculateEmissions()
 
-				const resCalc = new ResultCalculator()
-				resCalc.aggregate()
-				
-				const resultsColors = setResultsColorGroup()
-				speckleStore.setColorGroups(resultsColors)
-			} else if(newVal === 'Benchmark') {
-				console.log("Benchmark page - No graphics to update")
-			}
-		})
-
-		return {}
+		const resCalc = new ResultCalculator()
+		resCalc.aggregate()
+		
+		const resultsColors = setResultsColorGroup()
+		speckleStore.setColorGroups(resultsColors)
+	} else if(newVal === 'Benchmark') {
+		console.log("Benchmark page - No graphics to update")
 	}
-}
+})
 </script>
